license project
This is a site working with a MySQL database, which stores details about
movies such as: storyline, trailer,  and details about actors and
directors. Also an user can leave a comment or buy  the DVD of the movie
without having an account, when the user fills the fields about the
delivery an account will be created.